Fix #85: Work around a FPU bug in the x86 Windows on ARM Prims Emulation when assigning UInt64 to a Double.
The work around not only works around the Prism Emulator bug but the new code is even faster than the code that the Delphi compiler generates for the 32bit UInt64->Double conversion.

UInt64->Double conversion info
------------------------------
The Delphi 32bit compiler generates the following pseudo code:

  Int64Temp := Int64(UValue - 2^63);
  DoubleValue := ConvertInt64ToDouble(Int64Temp) + Double(2^63);

  mov eax,[UValue.Lo]
  mov edx,[UValue.Hi]
  mov [Int64Temp.Lo],eax
  sub edx,$80000000
  mov [Int64Temp.Hi],edx
  fild qword ptr [Int64Temp]
  fadd qword ptr [Double2ToThePowerOf63)]
  fstp qword ptr [DoubleValue]
  wait

The first line requires 8 additional Bytes on the stack for the temp variable, as UInt64/Int64 cannot be hold in a CPU register in 32bit. So we have 2 additional memory writes and 2 additional memory reads and an int64 subtraction.

The new code uses this pseudo code:

  if not IsSignBitSet(UValue) then
    DoubleValue := ConvertInt64ToDouble(Int64Temp)
  else
    DoubleValue := ConvertInt64ToDouble(Int64Temp) + Single(2^64);

  test byte ptr [UValue.Hi],$80
  jnz +8
  fild qword ptr [UValue]
  fstp qword ptr [DoubleValue]
  wait
  jmp +12
  fild qword ptr [UValue]
  fadd dword ptr [Single2ToThePowerOf64]
  fstp qword ptr [DoubleValue]
  wait

Now we have only 1 additional memory read and a conditional jump that is only taken for very large values. We don't need a temporary Int64 variable for the subtraction anymore and for "small" values we can skip the floating point addition. Furthermore the "fadd" with the Single(2^64) is the workaround for the ARM Prism Emulator ("fadd" resulted always in 0.0)